+++ src/renderer/pages/groups/groups.tsx
@@ -195,7 +195,7 @@ export function Groups(): JSX.Element {
               })}
             </div>
           ) : (
-            <div className="h-full w-full flex flex-col items-center text-center gap-4 justify-center text-lg">
+            <div className="h-full w-full text-justify gap-4 justify-center text-lg">
               <p className="text-black-600 dark:text-white font-medium">
                 {t('page.groups.createGroupText')}
               </p>